mysite/requirement.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
awscrt==0.9.10
awsiotsdk==1.5.2
beautifulsoup4==4.9.3
boto3==1.16.36
botocore==1.19.36
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-bootstrap4==2.3.1
django-storages==1.10.1
djangorestframework==3.12.2
gevent==20.9.0
greenlet==0.4.17
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.1.1
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.1
msgpack-python==0.5.6
mysqlclient==2.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
python-dateutil==2.8.1
pytz==2020.4
redis==2.10.6
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
wincertstore==0.2
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
